Update README.md (#165)
* Update README.md

includes link to AMIA 2020 presentation video
link to blog posts and AMIA 2020 presentation video in "How to use"
added dvmerge
add dvguidance

* Update README.md

I added in a link to https://mipops.github.io/dvrescue/  in dvguidance.

I also, revised the description for dvmerge based on Dave's comment. Is that what you meant or should it be moved to the dvrescue.xsd section?