refactor handling of timestamps with quicktime source
Disable the edit list and force the frame rate to match the encoded dv.
Add a check for coherency in first few dv frames in mov.
If the first frame is non-standard duration than find the offset to apply to subsequent pts references.
thx to @kieranjol for samples to find this