Update capture.md
This update includes reorganizing the capture instructions to include a section for CLI vs. GUI. The CLI instructions were also broken down to include a section explaining how to build a CLI string, examples of frequently used capture commands and a list of the full commands (the last of which still needs to be updated and examples need to be added).