feat: allow setting `dvplay`/`dvsampler` font via envvar
This commit allows the environment variable `DVRESCUE_FONT` to set the
font used by `dvplay` and `dvsampler`. This allows the font used to be
configured by the user, and also allows a font to be provided so that
the scripts can function properly when none of the hardcoded fonts could
be found successfully.

Related-to: gh-974